isolating fablib instantiation from token authentication class
